Move types to constants file
